+++ modules/desktop/hyprland/_features/main.nix
@@ -1,23 +1,6 @@
 # Compositor
-{
-  util ? null,
-  files ? null,
-  ...
-}: {
-  nixos = {
-    config,
-    lib,
-    pkgs,
-    ...
-  }: let
-    inherit
-      (lib)
-      getExe
-      getExe'
-      mkForce
-      replaceStrings
-      ;
-  in {
+_: {
+  nixos = {pkgs, ...}: {
     # WM
     programs.hyprland = {
       enable = true;
@@ -26,18 +9,6 @@
       package = pkgs.hyprworld.hyprland;
       portalPackage = pkgs.hyprworld.xdg-desktop-portal-hyprland;
     };
-
-    # Greeter
-    services.greetd.settings.default_session.command = mkForce "${getExe' config.programs.hyprland.package "start-hyprland"} -- --config ${
-      pkgs.writeText "greeter.conf" (
-        replaceStrings ["@greeter"] [(getExe config.programs.regreet.package)] (
-          util.build.theme {
-            inherit (config.lib.stylix) colors;
-            file = files.hyprland.greeter;
-          }
-        )
-      )
-    } &> /dev/null";
 
     # App Environment
     xdg.portal.extraPortals = [pkgs.xdg-desktop-portal-gtk];